latest sources

+++ TMessagesProj/src/main/java/org/cloudveil/messenger/api/model/request/SettingsRequest.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
  */
 
 public class SettingsRequest {
-    public int userId;
+    public long userId;
     public String userPhone;
     public String userName;
